% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nze_pollingplaces}
\alias{nze_pollingplaces}
\title{Polling place information.}
\format{A data frame:
\describe{
\item{\code{year}}{the election year}
\item{\code{district}}{the electoral district}
\item{\code{place}}{the polling place}
\item{\code{party}}{the party name}
\item{\code{mandate}}{the mandate type for which the votes have been recorded}
\item{\code{votes}}{the number of votes for the party at the polling place}
}}
\source{
\url{http://www.electionresults.govt.nz/}
}
\usage{
nze_pollingplaces
}
\description{
Electoral outcomes at the polling station level
}
\keyword{datasets}

