% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nze_votes}
\alias{nze_votes}
\title{Votes.}
\format{A data frame with seve variables: 
\describe{
\item{\code{year}}{the election year}
\item{\code{district}}{the electoral district}
\item{\code{party}}{the party name}
\item{\code{mandate}}{the type of mandate for which the votes have been recorded (cvote = Candidate Vote, pvote = Party Vote)}
\item{\code{votes}}{the number of votes}
\item{\code{per}}{the percentage of votes for the \code{party}}
}}
\source{
\url{http://www.electionresults.govt.nz/}
}
\usage{
nze_votes
}
\description{
Complete data on New Zealand federal elections from 1996 to 2014.
}
\keyword{datasets}

