% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nze_splitvotes}
\alias{nze_splitvotes}
\title{Joint and conditional distributions of candidate and party vote in New Zealand federal elections, 1996-2014}
\format{A data frame with nine variables:
\describe{
\item{\code{year}}{election year}
\item{\code{district}}{the electoral district}
\item{\code{cvote}}{the party of the candidate vote}
\item{\code{pvote}}{the party of the party vote} 
\item{\code{votes}}{the number of votes}
\item{\code{per_joint}}{the joint distribution (in \%) of party and candidate votes}
\item{\code{per_bar_cvote}}{the conditional distribution party vote|candidate vote (in \%)}
\item{\code{per_bar_pvote}}{the conditional distribution candidate vote|party vote (in \%)}
}}
\source{
\url{http://www.electionresults.govt.nz/}

For further details, see \url{http://www.electionresults.govt.nz/}
}
\usage{
nze_splitvotes
}
\description{
Joint distribution of candidate and party vote in New Zealand federal elections, 1996-2014
}
\keyword{datasets}

